v0.2.0
This release adds the `truncate` function, with takes some text, and an
argument for the line length, and will return a vector of lines from the
text, truncated to the specified line length. It accounts for words, and
so will not break a word across multiple lines. For words longer than
the line length specified (i.e. URLs), they will simply get pushed to
their own line.